+++ app/CHANGELOG.md
@@ -7,10 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-Update dependencies
+- Fixed period that does not start at the beginning of the day - [#22](https://trello.com/c/Og9bnQVr/22-bug-periodo)
 
-Fixed period that does not start at the beginning of the day
+- Add location to record form - [#11](https://trello.com/c/17BW9u8s/11-record-location)
 
 ## [1.0.0] - 2020-02-09
 
-Initial version
+- Initial version